added dyanamic dependent file installs

+++ src/run.sh
@@ -9,19 +9,21 @@ check() {
 
 imp() {
     echo "[WARNING] this will replace most of your rc files"
-    #read -p "continue? (Y/n)" DOG
+    read -p "continue? (Y/n)" DOG
 
-    #DOG=$(echo "$DOG" | tr '[:upper:]' '[:lower:]')
-    #[ ! "$DOG" = "y" ] &&  exit
+    DOG=$(echo "$DOG" | tr '[:upper:]' '[:lower:]')
+    [ ! "$DOG" = "y" ] &&  exit
 
     echo "importing started 🦑"
+    # create needed directories
+    [ ! -d $HOME/.config/neofetch ] && mkdir -p $HOME/.config/neofetch || echo neofetch directory already made 😘
 
-    mkdir -p $HOME/.config/neofetch
     cp $CONFILES/config.conf $HOME/.config/neofetch
+    cp $CONFILES/zsh-git-prompt.sh $HOME/.config
+
     cp $CONFILES/tmux.conf $HOME/.tmux.conf
     cp $CONFILES/vimrc $HOME/.vimrc
     cp $CONFILES/zshrc $HOME/.zshrc
-    cp $CONFILES/zsh-git-prompt.sh $HOME/.config
     
     echo "importing finished ✅"
 }
